+++ test/e2e/utils/networkpolicyspecbuilder.go
@@ -52,7 +52,7 @@ func (n *NetworkPolicySpecBuilder) SetName(namespace string, name string) *Netwo
 }
 
 // TODO: Add tests to match expressions
-func (n *NetworkPolicySpecBuilder) AddIngress(protoc v1.Protocol, port *int, portName *string, cidr *string, exceptCIDRs []string,
+func (n *NetworkPolicySpecBuilder) AddIngress(protoc v1.Protocol, port *int32, portName *string, cidr *string, exceptCIDRs []string,
 	podSelector map[string]string, nsSelector map[string]string,
 	podSelectorMatchExp []metav1.LabelSelectorRequirement, nsSelectorMatchExp []metav1.LabelSelectorRequirement) *NetworkPolicySpecBuilder {
 
@@ -116,7 +116,7 @@ func (n *NetworkPolicySpecBuilder) AddIngress(protoc v1.Protocol, port *int, por
 	if port != nil {
 		ports = []networkingv1.NetworkPolicyPort{
 			{
-				Port:     &intstr.IntOrString{IntVal: int32(*port)},
+				Port:     &intstr.IntOrString{IntVal: *port},
 				Protocol: &protoc,
 			},
 		}
@@ -152,7 +152,7 @@ func (n *NetworkPolicySpecBuilder) WithEgressDNS() *NetworkPolicySpecBuilder {
 	return n
 }
 
-func (n *NetworkPolicySpecBuilder) AddEgress(protoc v1.Protocol, port *int, portName *string, cidr *string, exceptCIDRs []string,
+func (n *NetworkPolicySpecBuilder) AddEgress(protoc v1.Protocol, port *int32, portName *string, cidr *string, exceptCIDRs []string,
 	podSelector map[string]string, nsSelector map[string]string,
 	podSelectorMatchExp []metav1.LabelSelectorRequirement, nsSelectorMatchExp []metav1.LabelSelectorRequirement) *NetworkPolicySpecBuilder {
